+++ checklist-app/src/main/java/com/datapath/checklistapp/dao/repository/DepartmentRepository.java
@@ -3,7 +3,7 @@ package com.datapath.checklistapp.dao.repository;
 import com.datapath.checklistapp.dao.entity.DepartmentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartmentRepository extends JpaRepository<DepartmentEntity, Long> {
+public interface DepartmentRepository extends JpaRepository<DepartmentEntity, Integer> {
 
-    DepartmentEntity findByIdentifier(Long identifier);
+    DepartmentEntity findByIdentifier(Integer identifier);
 }
