+++ checklist-app/src/main/java/com/datapath/checklistapp/dao/repository/QuestionGroupRepository.java
@@ -3,5 +3,5 @@ package com.datapath.checklistapp.dao.repository;
 import com.datapath.checklistapp.dao.entity.QuestionGroupEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionGroupRepository extends JpaRepository<QuestionGroupEntity, Long> {
+public interface QuestionGroupRepository extends JpaRepository<QuestionGroupEntity, Integer> {
 }
